Restructured project for tests
