--- back-end/AppServer/.idea/artifacts/AppServer_war.xml
@@ -1,13 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="war" name="AppServer:war">
-    <output-path>$PROJECT_DIR$/out/artifacts/AppServer_war</output-path>
-    <root id="archive" name="AppServer_war.war">
-      <element id="artifact" artifact-name="AppServer:war exploded" />
-      <element id="directory" name="WEB-INF">
-        <element id="directory" name="lib">
-          <element id="library" level="project" name="Jersey-2.2" />
-        </element>
-      </element>
-    </root>
-  </artifact>
-</component>
