--- back-end/AppServer/.idea/artifacts/AppServer_war_exploded.xml
@@ -1,13 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="exploded-war" name="AppServer:war exploded">
-    <output-path>$PROJECT_DIR$/out/artifacts/AppServer_war_exploded</output-path>
-    <root id="root">
-      <element id="javaee-facet-resources" facet="AppServer/web/Web" />
-      <element id="directory" name="WEB-INF">
-        <element id="directory" name="classes">
-          <element id="module-output" name="AppServer" />
-        </element>
-      </element>
-    </root>
-  </artifact>
-</component>
