--- back-end/AppServer/.idea/libraries/Jersey_2_2.xml
@@ -1,15 +1,0 @@
-<component name="libraryTable">
-  <library name="Jersey-2.2" type="JAX-RS">
-    <properties version="2.2" />
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/lib/jersey-client.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/jersey-container-servlet.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/javax.ws.rs-api-2.0.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/jersey-container-servlet-core.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/jersey-common.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/jersey-server.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
