--- back-end/AppServer/.idea/libraries/postgresql_42_1_1.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="postgresql-42.1.1">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/postgresql-42.1.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
