--- back-end/AppServer/.idea/libraries/postgresql_42_1_1__2_.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="postgresql-42.1.1 (2)">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/lib/postgresql-42.1.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
