+++ back-end/AppServer/src/main/java/com/simpleproject/ServeletFiles.java
@@ -1,3 +1,5 @@
+package main.java.com.simpleproject;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
